Refactor: ChannelManager 생성하여 채널셋업 설정
