+++ src/main/java/discord/podongbot/voice/BotVoiceControl.java
@@ -1,6 +1,5 @@
 package discord.podongbot.voice;
 
-import net.dv8tion.jda.api.audio.AudioSendHandler;
 import net.dv8tion.jda.api.entities.Guild;
 import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.entities.channel.concrete.VoiceChannel;
